Fully associative with hot/cold approx

+++ main.hpp
@@ -7,7 +7,7 @@
 #include <vector>
 #include <utility>
 
-typedef std::pair<bool, uint64_t> instr;
-const unsigned long MAX_TRACE_LENGTH = 1200000;
+typedef std::pair<bool, uint32_t> instr;
+const uint32_t MAX_TRACE_LENGTH = 1200000;
 
 #endif  // MAIN_HPP_
